Corrections nouvelle archi
git-svn-id: https://forge.indepnet.net/svn/glpi/glpi/trunk@3334 35612d4c-9d10-0410-837c-bf9dd1be60d7

+++ inc/common.class.php
@@ -311,7 +311,7 @@ class CommonDBTM {
 			$i++;
 		}		
 		for ($i=0; $i < count($fields); $i++) {
-			$query .= $fields[$i];
+			$query .= "`".$fields[$i]."`";
 			if ($i!=count($fields)-1) {
 				$query .= ",";
 			}
@@ -324,7 +324,7 @@ class CommonDBTM {
 			}
 		}
 		$query .= ")";
-
+		
 		if ($result=$db->query($query)) {
 			$this->post_addToDB();
 			return $db->insert_id();
@@ -555,7 +555,7 @@ class CommonDBTM {
 			if ($prev>0) echo "<li><a href='$cleantarget?ID=$prev'><img src=\"".$HTMLRel."pics/left.png\" alt='".$lang["buttons"][12]."' title='".$lang["buttons"][12]."'></a></li>";
 			if ($next>0) echo "<li><a href='$cleantarget?ID=$next'><img src=\"".$HTMLRel."pics/right.png\" alt='".$lang["buttons"][11]."' title='".$lang["buttons"][11]."'></a></li>";
 
-			if (function_exists("isReservable")&&isReservable(COMPUTER_TYPE,$ID)){
+			if (haveRight("reservation_central","r")&&function_exists("isReservable")&&isReservable(COMPUTER_TYPE,$ID)){
 				echo "<li class='invisible'>&nbsp;</li>";
 				echo "<li".(($actif==11)?" class='actif'":"")."><a href='$target&amp;onglet=11$template'>".$lang["title"][35]."</a></li>";
 			}
